chore: inline test in test file
